Featuer/box

See merge request s09-webmobile3-sub2/S09P12C104!26

+++ BACKEND/src/main/java/com/ssafy/keepham/domain/box/controller/SuperBoxController.java
@@ -2,10 +2,9 @@ package com.ssafy.keepham.domain.box.controller;
 
 import com.ssafy.keepham.common.api.Api;
 import com.ssafy.keepham.common.error.BoxError;
-import com.ssafy.keepham.domain.box.dto.BoxDTO;
 import com.ssafy.keepham.domain.box.dto.BoxRequest;
 import com.ssafy.keepham.domain.box.dto.BoxResponse;
-import com.ssafy.keepham.domain.box.entity.Box;
+import com.ssafy.keepham.domain.box.dto.BoxSaveRequest;
 import com.ssafy.keepham.domain.box.service.BoxService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -25,24 +24,23 @@ public class SuperBoxController {
 
     // 함생성
     @PostMapping
-    public Api<BoxResponse> createBox(@RequestBody BoxRequest boxRequest){
-
-        var res =  boxService.saveBox(boxRequest);
+    public Api<BoxResponse> createBox(@RequestBody BoxSaveRequest boxSaveRequest){
+        System.out.println("111"+boxSaveRequest.getDetailedAddress());
+        var res =  boxService.saveBox(boxSaveRequest);
         return Api.OK(res);
     }
 
     //특정 id의 함 조회
     @GetMapping("/{boxId}")
     public Api<Object> getBox(@PathVariable Long boxId){
+
         if(boxId<=0){
             return Api.ERROR(BoxError.BOX_BAD_REQUEST, String.format("[%d]은/는 요휴하지 않는 id형식 입니다. 1이상의 숫자로 요청해 주세요.", boxId));
         }
 
         var res = boxService.getBoxById(boxId);
-        if(res == null){
-            return Api.ERROR(BoxError.BOX_NOT_FOUND, String.format("[%d]은/는 존재하지 않는 함의 id입니다.", boxId));
-        }
-        else if(res.isValid()){
+
+        if(res.isValid()){
             return Api.OK(res);
         }
         else {
@@ -53,24 +51,33 @@ public class SuperBoxController {
 
     //삭제로 변환 안된 box들 조회
     @GetMapping
-    public List<BoxDTO> getAllBox(){
-        return boxService.getAllBox();
+    public Api<List<BoxResponse>> getAllBox(){
+        return  Api.OK(boxService.getAllBox());
     }
 
     //박스 수정
     @PutMapping("/{boxId}")
-    public Box updateBox(@PathVariable Long boxId, @RequestBody Box box){
-        box.setId(boxId);
-        return boxService.updateBox(box);
-    }
+    public  Api<Object> updateBox(@PathVariable Long boxId, @RequestBody BoxRequest boxRequest){
+        if(boxId<=0){
+            return Api.ERROR(BoxError.BOX_BAD_REQUEST, String.format("[%d]은/는 요휴하지 않는 id형식 입니다. 1이상의 숫자로 요청해 주세요.", boxId));
+        }
+        var res  = boxService.updateBox(boxId, boxRequest);
+        return Api.OK(res);
+
 
     //박스 삭제 상태로 전환
     @PutMapping("/delete/{boxId}")
-    public Box deleteBox(@PathVariable Long boxId){
-        return boxService.deleteBox(boxId);
+    public Api<Object> deleteBox(@PathVariable Long boxId){
+
+        if(boxId<=0){
+            return Api.ERROR(BoxError.BOX_BAD_REQUEST, String.format("[%d]은/는 요휴하지 않는 id형식 입니다. 1이상의 숫자로 요청해 주세요.", boxId));
+        }
+
+        var res = boxService.deleteBox(boxId);
+
+        return Api.OK(res);
+
     }
-
-
